mv mathjax axis title test to a bundle_test suite
- Removing MathJax from the page currently leads to errors
  in downstream tests.
